refactor: 레이싱카 패키지명 변경
