+++ src/main/java/racingcar/CurrentLocation.java
@@ -1,4 +1,4 @@
-package step03;
+package racingcar;
 
 public class CurrentLocation {
