+++ src/main/java/racingcar/OutputView.java
@@ -1,4 +1,4 @@
-package step03;
+package racingcar;
 
 import java.util.List;
