+++ src/main/java/racingcar/RacingCarApplication.java
@@ -1,4 +1,4 @@
-package step03;
+package racingcar;
 
 import java.util.List;
 import java.util.stream.Collectors;
