+++ src/test/java/racingcar/RandomNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package step03;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
